reorganize dementia disease model, cleanup comments
