+++ src/vivarium_experiment_dementia/components/__init__.py
@@ -1,3 +1,2 @@
 from .cdr import DementiaProgression
-from .dementia_state import DementiaExcessMortalityState
 from .dementia_SI import Dementia_SI
