+++ src/vivarium_experiment_dementia/components/cdr.py
@@ -1,13 +1,11 @@
 import pandas as pd
 
-# TODO: don't waffle between cdr_sb and cdr. pick one
-
 
 class DementiaProgression:
 
     configuration_defaults = {
         'dementia_model': {
-            'cdr_sb_rate': 2.44  # annual
+            'cdr_sb_rate': 2.44  # annual change
         }
     }
 
@@ -21,7 +19,7 @@ class DementiaProgression:
         builder.population.initializes_simulants(self.on_initialize_simulants, creates_columns=['cdr'],
                                                  requires_columns=['alzheimers_disease_and_other_dementias'])
 
-        # register cdr_sb_rate value producer to give a hook for treatment to reach in with
+        # register cdr_rate value producer to give a hook for treatment to reach in with
         self.cdr_rate = builder.value.register_rate_producer('cdr_rate', self._cdr_rate)
 
         self.pop_view = builder.population.get_view(['alive', 'alzheimers_disease_and_other_dementias', 'cdr'])
@@ -31,7 +29,7 @@ class DementiaProgression:
 
         builder.event.register_listener('time_step', self.on_time_step)
 
-    # TODO: Not everyone starts right at the beginning. Use prevalence or something to mix people around
+    # TODO: Not everyone starts right at the beginning of the disease. Use prevalence or something to mix people around
     def on_initialize_simulants(self, pop_data):
         pop = self.pop_subview.get(pop_data.index)
         pop['cdr'] = 0.0
@@ -47,8 +45,7 @@ class DementiaProgression:
         self.pop_view.update(pop)
 
     def _cdr_rate(self, index):
-        # zero if don't have, otherwise default.
-        # the treatment component will modify this as needed
+        # zero if don't have, otherwise default. The treatment component will modify this as needed
         pop = self.pop_view.get(index)
 
         cdr_rate = pd.Series(0.0, index=index)
